Scene Change Code Implemented
The intro scene now changes to the main when you press space bar.

But there's a weird lighting issue where changing the lighting in one
scene messes with the lighting in the other scene? Might be because
they're a copy of each other, so I'll next try deleting the second light
object and making a clean one.